% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFDR.r
\name{BayesianFDR}
\alias{BayesianFDR}
\title{Bayesian FDR}
\usage{
BayesianFDR(belief)
}
\arguments{
\item{belief}{A posterior belief vector.}
}
\value{
Bayesian q-value, which is analog to frequentists' q-value
}
\description{
Baysian FDR
}
\details{
This function is used to calculate Baysian FDR.
The result is similar to q-value.
}
\examples{
set.seed(15213)
aBeliefVec <- runif(1000)
bayesianQValue <- BayesianFDR(aBeliefVec)
}
\author{
Zhiguang Huo <zhuo@ufl.edu>
}
