% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMP_EB.r
\name{BayesMP_EB}
\alias{BayesMP_EB}
\title{MCMC for BayesMP_EB}
\usage{
BayesMP_EB(Z, gamma = NULL, updateGamma = TRUE, beta = 1/2,
  empMu = rep(0, ncol(Z)), empSD = rep(1, ncol(Z)), niter = 100,
  burnin = 50, silence = F, logDotsPerLine = 50,
  fileName = "BayesMP_EB_", writeY = F, writePi = F, writeGamma = F,
  writeMu = F, writeS2 = F, writeHSall = T)
}
\arguments{
\item{Z}{Z statistics. Z should be a p*n matrix.}

\item{gamma}{initial gamma Estimated null proportation.}

\item{updateGamma}{update gamma if true.}

\item{beta}{Non-informative prior: given a gene is DE, the prior probablity this gene is up-regulated, default=1/2}

\item{empMu}{a vector of mean parameter for the null component, default is 0.}

\item{empSD}{a vector of sd parameter for the null component, default is 1.}

\item{niter}{Number of iterations. Default 100, suggest to be 10,000}

\item{burnin}{Number of burnin period. Default 50, suggest to be 500}

\item{fileName}{Base fileName for saving fulll mcmc results, hypothesis HS results.}

\item{alpha}{Concentration parameter for DPs, default=1}

\item{mu0}{Mean parameter for base function, default=0}

\item{sigma0}{sqrt root of variance parameter for base function, default=10}

\item{sigma}{sqrt root of variance parameter for DP mixture component, default=1}

\item{trunc}{truncation parameter for base function (For both positive component and negative component), default=0}

\item{Pi}{initial pi (DE probablity vector), default rbeta(G, estGamma/(G-estGamma), 1)}

\item{delta}{initial delta (DE direction probablity vector), default rbeta(G, beta, beta)}

\item{Y}{initial component indicator. Y should be a p*n matrix. Y = ..., -2, -1, 0, 1, 2, ...}

\item{fullRes}{binary: 0: do not save full mcmc results; 1: save full mcmc results.}

\item{HSall}{binary: 0: do not save hypothesis HS results; 1: save hypothesis HS results.}
}
\value{
computing time
}
\description{
BayesMP_EB
}
\details{
implementation for BayesMP, MCMC part. This is a full Bayesian model, and the alternative distribution is modeled via dirichlet process.
}
\examples{
G <- 5000
K <- 10
alpha <- 200
X0 <- matrix(rnorm(G * K), G, K)
X1 <- matrix(rnorm(G * K, 2), G, K)
X2 <- matrix(rnorm(G * K, -2), G, K)
piall <- rbeta(G, alpha/G, 1)
delta <- rbeta(G, 1/2, 1/2)
p0 <- 1 - piall
p1 <- piall * delta
p2 <- piall * (1 - delta)
Y <- replicate(K, rMultinom(cbind(p0, p1, p2)))
Z <- X0 * (Y == 1) + X1 * (Y == 2) + X2 * (Y == 3)
niter <- 2000
system.time(BayesMP_EB(Z,fullRes=1))
}
\author{
Zhiguang Huo
}
