% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(plus, minus, nsample, alpha = 0, method = "cos")
}
\arguments{
\item{plus}{p*n matrix. Each element represents the posterior probablity of a gene to be positive DE.}

\item{minus}{p*n matrix. Each element represents the posterior probablity of a gene to be negative DE.}

\item{nsample}{Total number of posterior samples excluding burning period.}

\item{alpha}{Forge parameter}

\item{method}{Choose one from those: cos, KL, L1, L2, L2_2D, Hellinger.}
}
\value{
Dissimilarity matrix. p*p
}
\description{
Dissimilarity matrix calculation.
}
\details{
Based on DE posterior probablity of gene*study matrix,
calculate gene-gene distance
}
\examples{
G<-10
S<-3
nsample <- 10000
plus <- matrix(c(sample(nsample,G*S,replace = TRUE),rep(0,G*S)),2*G,S)
minus <- matrix(c(rep(0,G*S),sample(nsample,G*S,replace = TRUE)),2*G,S)
dissimilarity <- distance(plus, minus, nsample)
}
\author{
Zhiguang Huo
}

