% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNull_Efron_delta.r
\name{estimateNull_Efron_delta}
\alias{estimateNull_Efron_delta}
\title{Estimate null proportion in single study.}
\usage{
estimateNull_Efron_delta(x, delta)
}
\arguments{
\item{x}{Vector of Z statistics (transformed p value) in single study.}

\item{delta}{a neighborhood (-delta, delta) around 0 to be used to estiamte null proportion.}
}
\value{
Null proportion in single study.
}
\description{
The proportion of null component is estimated following the procedure by Storey (2002).
}
\details{
Estimate null proportion in single study.
}
\examples{
set.seed(15213)
n1 <- 1000
n2 <- 100
aZ <- c(rnorm(n1),runif(n2,2,3)*sample(n2))
delta <- 0.4
estimateNull_Efron_delta(aZ, delta)
}
\author{
Zhiguang Huo
}

