% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePrior.r
\name{estimatePrior}
\alias{estimatePrior}
\title{Estimate null proportion for multiple studies.}
\usage{
estimatePrior(X, delta)
}
\arguments{
\item{X}{Matrix of Z statistics (transformed p value) p*S, p is number of genes and S is number of studies.}

\item{delta}{a neighborhood (-delta, delta) around 0 to be used to estiamte null proportion.}
}
\value{
Null proportion estimation of multiple studies.
}
\description{
The proportion of null component is estimated following the procedure by Storey (2002).
}
\details{
Estimate null proportion for multiple studies.
}
\examples{
n1 <- 1000
n2 <- 100
S <- 3
set.seed(15213)
aZ <- c(rnorm(n1),rnorm(n2,3)*sample(n2))
set.seed(15214)
bZ <- c(rnorm(n1),rnorm(n2,3)*sample(n2))
set.seed(15215)
cZ <- c(rnorm(n1),rnorm(n2,3)*sample(n2))
Z <- cbind(aZ, bZ, cZ)
delta <- 0.4
estimateNull_Efron_delta(Z, delta)
}
\author{
Zhiguang Huo
}

